requirements.txt
Pillow
qt4w
chrome_master
qtaf